# ENH: Add vectorized implementation for stochastic optimizers in MLPPerf optimizers