--- a/sklearn/preprocessing/_data.py
+++ b/sklearn/preprocessing/_data.py
@@ -4,7 +4,6 @@
 
 import warnings
 from numbers import Integral, Real
-
 import numpy as np
 from scipy import sparse, stats
 from scipy.special import boxcox, inv_boxcox
@@ -229,7 +228,9 @@ def scale(X, *, axis=0, with_mean=True, with_std=True, copy=True):
         estimator="the scale function",
         dtype=FLOAT_DTYPES,
         ensure_all_finite="allow-nan",
-    )
+        input_name="X",
+)
+
     if sparse.issparse(X):
         if with_mean:
             raise ValueError(