### Summary
This PR adds a vectorized implementation of the stochastic optimizers used in MLP (`_stochastic_optimizers.py`). The goal is to improve performance by leveraging NumPy vectorized operations instead of explicit Python loops.

### Changes
- Added `_stochastic_optimizers_vec.py` with vectorized code.
- `_stochastic_optimizers.py` updated to use the vectorized version.
- Original implementation saved as `_stochastic_optimizers_orig.py`.

### Performance
Benchmarks on synthetic datasets (20,000 samples, 50 features):

| Task                     | Original (s) | Vectorized (s) |
|--------------------------|--------------|----------------|
| Classification (SGD)     | 9.38         | 10.96          |
| Classification (Adam)    | 11.53        | 11.82          |
| Regression (SGD)         | 8.94         | 9.69           |
| Regression (Adam)        | 10.03        | 8.79           |

### Notes
- All tests pass.
- Benchmark scripts (`benchmark_optimizers.py`) and results are **not included** in this PR.
<img width="800" height="500" alt="benchmark_comparison" src="https://github.com/user-attachments/assets/6b4bd984-bfbc-4967-9142-7512b35b869e" />